@@ -2,6 +2,7 @@
   "name": "web",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev --turbopack",
     "build": "next build",
@@ -17,28 +18,26 @@
     "@azure/msal-browser": "3.24.0",
     "@azure/msal-react": "2.1.0",
     "@azure/storage-blob": "12.17.0",
-    "@chakra-ui/icons": "2.1.1",
-    "@chakra-ui/react": "2.5.5",
-    "@emotion/react": "11.10.5",
-    "@emotion/styled": "11.10.5",
+    "@chakra-ui/react": "3.8.0",
+    "@emotion/react": "^11.14.0",
     "@fontsource/hanken-grotesk": "4.5.2",
     "@hey-api/client-axios": "0.6.1",
     "@tanstack/react-query": "5.0.5",
     "axios": "1.5.0",
-    "chakra-react-select": "4.7.6",
+    "chakra-react-select": "6.0.1",
     "fastest-levenshtein": "1.0.16",
     "form-data": "4.x",
-    "framer-motion": "10.16.5",
+    "framer-motion": "12.4.1",
     "jwt-decode": "4.0.0",
-    "next": "15.0.3",
+    "next": "15.2.3",
+    "next-themes": "^0.4.4",
     "pdf-lib": "1.17.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-draggable": "4.4.6",
     "react-dropzone": "14.3.5",
-    "react-icons": "5.3.0",
+    "react-icons": "^5.4.0",
     "react-pdf": "7.7.0",
-    "react-query": "3.39.3",
     "react-rnd": "10.4.13",
     "react-router-dom": "6.8.1",
     "react-scripts": "5.0.1",
