@@ -8,14 +8,14 @@ import {
 } from '@nestjs/common';
 import { AssignedGroupService } from './assigned-group.service';
 import { UpdateAssignedGroupSignerDto } from './dto/update-assigned-group-signer.dto';
-import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { ContributorAuthGuard } from '../auth/guards/contributor-auth.guard';
 
 @Controller('signatures')
 export class AssignedGroupController {
   constructor(private readonly assignedGroupService: AssignedGroupService) {}
 
   @Patch(':id/signer')
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(ContributorAuthGuard)
   updateAssignedGroupSigner(
     @Param('id') assignedGroupId: string,
     @Body(new ValidationPipe({ transform: true }))
