@@ -1,5 +1,11 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { ArrayMinSize, IsArray, IsNotEmpty, IsString } from 'class-validator';
+import {
+  ArrayMinSize,
+  IsArray,
+  IsNotEmpty,
+  IsOptional,
+  IsString,
+} from 'class-validator';
 import { CreateAssignedGroupDto } from '../../assigned-group/dto/create-assigned-group.dto';
 
 export class CreateFormInstanceDto {
@@ -7,6 +13,11 @@ export class CreateFormInstanceDto {
   @IsNotEmpty()
   @ApiProperty()
   name: string;
+
+  @IsString()
+  @ApiProperty()
+  @IsOptional()
+  description: string;
 
   @IsArray()
   @ArrayMinSize(1)
