@@ -1,15 +1,13 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../prisma/prisma.service';
-import { ServerClient } from 'postmark';
+import { Inject, Injectable } from '@nestjs/common';
 import { Employee } from '@prisma/client';
+import { EmailHandler } from './EmailHandlerInterface';
 
 @Injectable()
 export class PostmarkService {
-  client: ServerClient;
-
-  constructor(private prisma: PrismaService) {
-    this.client = new ServerClient(process.env.POSTMARK_SERVER_KEY as string);
-  }
+  constructor(
+    @Inject('EmailHandler')
+    private emailHandler: EmailHandler,
+  ) {}
 
   /**
    * Send an email using Postmark
@@ -19,22 +17,7 @@ export class PostmarkService {
    * @returns Promise<void>
    */
   async sendEmail(to: string, subject: string, textBody: string) {
-    if (process.env.NODE_ENV === 'development') {
-      console.log('Email:', { to, subject, textBody });
-      return;
-    }
-
-    try {
-      this.client.sendEmail({
-        From: 'jfrederick@mfa.org',
-        To: to,
-        Subject: subject,
-        TextBody: textBody,
-        MessageStream: 'outbound',
-      });
-    } catch (error) {
-      console.error('Error sending email through postmark:', error);
-    }
+    await this.emailHandler.sendEmail(to, subject, textBody);
   }
 
   /**
@@ -48,10 +31,7 @@ export class PostmarkService {
     originatorName: string,
     formName: string,
   ) {
-    const subject: string = `Hi ${originatorName}, you have created a new form: ${formName}.`;
-    const body: string = `Hi ${originatorName}, you have created a new form: ${formName}.`;
-
-    await this.sendEmail(to, subject, body);
+    await this.emailHandler.sendFormCreatedEmail(to, originatorName, formName);
   }
 
   /**
@@ -65,10 +45,11 @@ export class PostmarkService {
     signerName: string,
     formName: string,
   ) {
-    const subject: string = `Form ${formName} Ready To Sign`;
-    const body: string = `Hi ${signerName}, you have a form ready for your signature: ${formName}.`;
-
-    await this.sendEmail(to, subject, body);
+    await this.emailHandler.sendReadyForSignatureToUserEmail(
+      to,
+      signerName,
+      formName,
+    );
   }
 
   /**
@@ -80,21 +61,10 @@ export class PostmarkService {
     departmentId: string,
     formName: string,
   ) {
-    const employees = await this.prisma.employee.findMany({
-      where: {
-        position: {
-          departmentId,
-        },
-      },
-    });
-
-    employees.forEach(async (employee) => {
-      await this.sendReadyForSignatureToUserEmail(
-        employee.email,
-        `${employee.firstName} ${employee.lastName}`,
-        formName,
-      );
-    });
+    await this.emailHandler.sendReadyForSignatureToDepartmentEmail(
+      departmentId,
+      formName,
+    );
   }
 
   /**
@@ -106,19 +76,10 @@ export class PostmarkService {
     positionId: string,
     formName: string,
   ) {
-    const employees = await this.prisma.employee.findMany({
-      where: {
-        positionId,
-      },
-    });
-
-    employees.forEach(async (employee) => {
-      await this.sendReadyForSignatureToUserEmail(
-        employee.email,
-        `${employee.firstName} ${employee.lastName}`,
-        formName,
-      );
-    });
+    await this.emailHandler.sendReadyForSignatureToPositionEmail(
+      positionId,
+      formName,
+    );
   }
 
   /**
@@ -130,13 +91,10 @@ export class PostmarkService {
     userList: Employee[],
     formName: string,
   ) {
-    userList.forEach(async (user) => {
-      await this.sendReadyForSignatureToUserEmail(
-        user.email,
-        `${user.firstName} ${user.lastName}`,
-        formName,
-      );
-    });
+    await this.emailHandler.sendReadyForSignatureToUserListEmail(
+      userList,
+      formName,
+    );
   }
 
   /**
@@ -153,10 +111,12 @@ export class PostmarkService {
     signerName: string,
     formName: string,
   ) {
-    const subject: string = `Form ${formName} Signed By ${signerName}`;
-    const body: string = `Hi ${originatorName}, your form ${formName} has been signed by user: ${signerName}.`;
-
-    await this.sendEmail(to, subject, body);
+    await this.emailHandler.sendSignedEmail(
+      to,
+      originatorName,
+      signerName,
+      formName,
+    );
   }
 
   /**
@@ -170,9 +130,10 @@ export class PostmarkService {
     originatorName: string,
     formName: string,
   ) {
-    const subject: string = `Form ${formName} Ready for Approval`;
-    const body: string = `Hi ${originatorName}, your form ${formName} is completed and is ready for your approval: ${formName}.`;
-
-    await this.sendEmail(to, subject, body);
+    await this.emailHandler.sendReadyForApprovalEmail(
+      to,
+      originatorName,
+      formName,
+    );
   }
 }
