@@ -1,10 +1,9 @@
-import { HStack, Flex, Text, useDisclosure } from '@chakra-ui/react';
-import { FormCard } from './FormCard';
-import { FormInstanceEntity } from '@web/client/types.gen';
+import { Flex, Text, useDisclosure, Wrap } from '@chakra-ui/react';
+import { FormCard } from './FormCard.tsx';
+import { FormInstanceEntity } from '@web/client/types.gen.ts';
 import React, { useState } from 'react';
 import { FormImageCard } from './FormImageCard';
 import { ViewAll } from './ViewAll';
-import { ProfileIcon } from '@web/static/icons';
 import { SignFormInstancePreview } from './SignFormInstancePreview';
 
 /**
@@ -30,12 +29,12 @@ export const OverviewRow = ({
     0,
     Math.min(4, formInstances.length),
   );
-  const { isOpen, onOpen, onClose } = useDisclosure();
+  const [isOpen, setIsOpen] = useState(false);
   const [curForm, setCurForm] = useState<FormInstanceEntity>();
 
   function handleModalOpen(formInstance: FormInstanceEntity) {
     setCurForm(formInstance);
-    onOpen();
+    setIsOpen(true);
   }
 
   return (
@@ -70,12 +69,7 @@ export const OverviewRow = ({
           <ViewAll title={title} link={link} />
         </Flex>
       </Flex>
-      <Flex
-        marginTop="20px"
-        flexDirection="row"
-        width="100%"
-        justifyContent={'space-between'}
-      >
+      <Wrap gap="20px" columnGap="20px" justify="flex-start">
         {displayFormInstances.map(
           (formInstance: FormInstanceEntity, index: number) => {
             return title == 'To-do' ? (
@@ -95,10 +89,10 @@ export const OverviewRow = ({
             );
           },
         )}
-      </Flex>
+      </Wrap>
       <SignFormInstancePreview
         isOpen={isOpen}
-        onClose={onClose}
+        onClose={() => setIsOpen(false)}
         formInstance={curForm}
       />
     </>
