@@ -1,18 +1,11 @@
-import {
-  Avatar,
-  Button,
-  Flex,
-  Modal,
-  ModalContent,
-  ModalOverlay,
-  Text,
-} from '@chakra-ui/react';
-import { FormInstanceEntity } from '@web/client/types.gen';
-
+import { Button, Dialog, Flex, Portal, Text } from '@chakra-ui/react';
+import { AssignedGroupEntity, FormInstanceEntity } from '@web/client/types.gen';
 import { useRouter } from 'next/router';
 import { CloseIcon, PenSigningIcon } from '@web/static/icons';
-import AssigneeMap from './AvatarMap';
 import { getNameFromAssignedGroup } from '@web/utils/formInstanceUtils';
+import { useAuth } from '@web/hooks/useAuth';
+import AssigneeMap from './AssigneeMap';
+import { Avatar } from './ui/avatar.tsx';
 
 /**
  * Modal used in OverviewRow component for To Do forms
@@ -27,10 +20,11 @@ export const SignFormInstancePreview = ({
   formInstance,
 }: {
   isOpen: boolean;
-  onClose: any;
+  onClose: () => void;
   formInstance?: FormInstanceEntity;
 }) => {
   const router = useRouter();
+  const { user } = useAuth();
 
   const subheadingStyle = {
     lineHeight: 'normal',
@@ -40,149 +34,183 @@ export const SignFormInstancePreview = ({
   };
 
   if (!formInstance) {
-    return null;
+    return <></>;
+  }
+
+  /**
+   * Determines whether the currently logged in user qualifies as 'next to sign' based on
+   * their position/department/identity.  Used to determine whether the user should have the option to
+   * "Sign Now" from this form instance preview modal.
+   * @returns a boolean representing whether or not the user can sign for the next assigned group
+   */
+  function nextUser() {
+    if (!formInstance) {
+      return false;
+    }
+    // get the AssignedGroupEntity that is next to sign the form
+    let nextToSign: AssignedGroupEntity | null = null;
+    let prev: AssignedGroupEntity | null = null;
+    for (let assigned of formInstance.assignedGroups) {
+      if (prev == null || (prev.signed && !assigned.signed)) {
+        nextToSign = assigned;
+        break;
+      }
+      prev = assigned;
+    }
+    if (nextToSign?.signerType == 'POSITION') {
+      return user?.positionId == nextToSign.signerPositionId;
+    } else if (nextToSign?.signerType == 'DEPARTMENT') {
+      return user?.departmentId == nextToSign.signerDepartmentId;
+    } else if (nextToSign?.signerType == 'USER') {
+      return user?.id == nextToSign.signerEmployee?.id;
+    } else if (nextToSign?.signerType == 'USER_LIST') {
+      return nextToSign.signerEmployeeList?.reduce(
+        (acc, empl) => (empl.id == user?.id ? true : acc),
+        false,
+      );
+    }
+    return false;
   }
 
   return (
-    <Modal
-      isOpen={isOpen}
-      onClose={onClose}
-      isCentered
-      closeOnOverlayClick={false}
+    <Dialog.Root
+      open={isOpen}
+      onOpenChange={onClose}
+      closeOnInteractOutside={true}
     >
-      <ModalOverlay bg="rgba(0, 0, 0, 0.5)" />
-      <ModalContent alignItems="center" justifyContent={'center'}>
-        <Flex
-          backgroundColor="#F8F9FA"
-          padding="24px 32px"
-          gap="24px"
-          flexDirection="column"
-          justifyContent={'space-between'}
-          width="559px"
-          height="554px"
-          borderRadius="12px"
-          alignItems="flex-end"
-          flexShrink={'0'}
-          box-shadow="0px 2px 16px 0px rgba(0, 0, 0, 0.15)"
-        >
-          <Flex
-            width="495px"
-            justifyContent="space-between"
-            alignItems="center"
-            flexDirection="column"
+      <Portal>
+        <Dialog.Backdrop bg="rgba(0, 0, 0, 0.5)" />
+        <Dialog.Positioner alignItems="center" justifyContent={'center'}>
+          <Dialog.Content
+            padding={'24px 32px'}
+            gap="24px"
+            backgroundColor="#F8F9FA"
+            flexDir={'column'}
+            width="559px"
+            minHeight="554px"
+            maxHeight="75vh"
+            borderRadius="12px"
+            boxShadow="0px 2px 16px 0px rgba(0, 0, 0, 0.15)"
+            overflowY={'scroll'}
           >
-            <Flex
-              flexDirection="column"
-              alignItems="flex-start"
-              gap="24px"
-              flex="1 0 0"
-            >
+            <Dialog.Header>
               <Flex
+                width="100%"
+                flexDirection="row"
                 justifyContent={'space-between'}
-                alignItems="center"
-                width="495px"
+                alignItems={'center'}
               >
-                <Text
-                  fontFamily="Hanken Grotesk"
+                <Dialog.Title
+                  fontFamily={'Hanken Grotesk'}
                   fontSize="19px"
-                  fontWeight="700px"
+                  fontWeight="19px"
                   lineHeight="26px"
                 >
                   {formInstance.name}
-                </Text>
-                <CloseIcon onClick={onClose} cursor="pointer" />
-              </Flex>
-
-              <Flex
-                flexDirection="column"
-                alignItems="flex-start"
-                gap="8px"
-                alignSelf="stretch"
-              >
-                <Text style={subheadingStyle}>Description</Text>
-                <Text color="#222324" fontSize="16px" fontWeight="400">
-                  For HR needs lorem ipsum dolor sit amet. Consectetur
-                  adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-                  dolore magna aliqua.
-                </Text>
-              </Flex>
-              <Flex
-                flexDirection="column"
-                alignItems="flex-start"
-                gap="12px"
-                alignSelf="stretch"
-              >
-                <Text style={subheadingStyle}>Assigned by</Text>
-                <Flex alignItems="center" gap="8px" alignSelf="stretch">
-                  <Avatar
-                    name={
-                      formInstance.originator.firstName +
-                      ' ' +
-                      formInstance.originator.lastName
-                    }
-                    boxSize="32px"
-                    backgroundColor={'#DDD'}
-                    padding="6px 7px"
-                    border="1px solid #FFFFFF"
-                    color="#0C0C0C"
-                    size="16px"
-                  />
-                  <Text color="#0C0C0C" fontSize="15px">
-                    {formInstance.originator.firstName +
-                      ' ' +
-                      formInstance.originator.lastName}
-                  </Text>
-                </Flex>
-              </Flex>
-              <Flex
-                flexDirection="column"
-                justifyContent={'space-between'}
-                alignSelf={'stretch'}
-                gap="12px"
-                overflowY={'scroll'}
-              >
-                <Text style={subheadingStyle}>Assignees</Text>
-                <AssigneeMap
-                  assignees={formInstance.assignedGroups.map(
-                    (assignedGroup) => ({
-                      signed: assignedGroup.signed,
-                      title: getNameFromAssignedGroup(assignedGroup),
-                      signerType: assignedGroup.signerType as any,
-                      updatedAt: assignedGroup.updatedAt,
-                    }),
-                  )}
+                </Dialog.Title>
+                <CloseIcon
+                  onClick={onClose}
+                  cursor="pointer"
+                  style={{
+                    width: '19px',
+                    height: '19px',
+                  }}
                 />
               </Flex>
-            </Flex>
-          </Flex>
+            </Dialog.Header>
+            <Dialog.Body>
+              <Flex flexDirection="column" alignItems="flex-start" gap="24px">
+                <Flex
+                  flexDirection="column"
+                  alignItems="flex-start"
+                  gap="8px"
+                  alignSelf="stretch"
+                >
+                  <Text style={subheadingStyle}>Description</Text>
+                  <Text color="#222324" fontSize="16px" fontWeight="400">
+                    {formInstance.description}
+                  </Text>
+                </Flex>
+                <Flex
+                  flexDirection="column"
+                  justifyContent={'space-between'}
+                  alignItems="flex-start"
+                  alignSelf="stretch"
+                  gap="12px"
+                >
+                  <Text style={subheadingStyle}>Assigned by</Text>
+                  <Flex alignItems={'center'}>
+                    <Avatar
+                      name={
+                        formInstance.originator.firstName +
+                        ' ' +
+                        formInstance.originator.lastName
+                      }
+                      boxSize="32px"
+                      padding="6px 7px"
+                      size="md"
+                      color="#0C0C0C"
+                      bg="#E5E5E5"
+                      border="1px solid #FFFFFF"
+                    />
+                    <Text
+                      color="#0C0C0C"
+                      fontSize="15px"
+                      lineHeight="20px"
+                      paddingLeft="8px"
+                      fontWeight="400"
+                    >
+                      {formInstance.originator.firstName}{' '}
+                      {formInstance.originator.lastName}
+                    </Text>
+                  </Flex>
+                </Flex>
+                <Flex
+                  flexDirection="column"
+                  justifyContent={'space-between'}
+                  alignSelf={'stretch'}
+                  gap="12px"
+                >
+                  <Text style={subheadingStyle}>Assignees</Text>
+                  <AssigneeMap
+                    assignees={formInstance.assignedGroups.map(
+                      (assignedGroup) => ({
+                        signed: assignedGroup.signed,
+                        title: getNameFromAssignedGroup(assignedGroup),
+                        signerType: assignedGroup.signerType as any,
+                        updatedAt: assignedGroup.updatedAt,
+                      }),
+                    )}
+                  />
+                </Flex>
+              </Flex>
+            </Dialog.Body>
+            <Dialog.Footer>
+              {nextUser() && (
+                <Button
+                  width="158px"
+                  height="32px"
+                  padding="4px 16px"
+                  borderRadius="6px"
+                  background="#1367EA"
+                  onClick={() =>
+                    router.push('form-instances/' + formInstance.id)
+                  }
+                  _hover={{
+                    background: '#1367EA',
+                  }}
+                >
+                  <Flex gap="8px" alignItems="center" justifyContent="center">
+                    <PenSigningIcon color="#FFF" />
 
-          <Button
-            width="158px"
-            height="32px"
-            padding="4px 16px"
-            borderRadius="6px"
-            background="#1367EA"
-            onClick={() => router.push('form-instances/' + formInstance.id)}
-            _hover={{
-              background: '#1367EA',
-            }}
-            justifyContent="center"
-            alignItems={'center'}
-          >
-            <Flex gap="8px">
-              <PenSigningIcon
-                color="#FFF"
-                position={'absolute'}
-                alignSelf="center"
-                left="20px"
-              />
-              <Text color="#FFF" position={'relative'}>
-                Sign Now
-              </Text>
-            </Flex>
-          </Button>
-        </Flex>
-      </ModalContent>
-    </Modal>
+                    <Text color="#FFF">Sign Now</Text>
+                  </Flex>
+                </Button>
+              )}
+            </Dialog.Footer>
+          </Dialog.Content>
+        </Dialog.Positioner>
+      </Portal>
+    </Dialog.Root>
   );
 };
