@@ -2,7 +2,7 @@ import { Box, Text, Button } from '@chakra-ui/react';
 import {
   PageSwitcherArrowLeft,
   PageSwitcherArrowRight,
-} from '@web/static/icons';
+} from '@web/static/icons.tsx';
 import { Dispatch, SetStateAction } from 'react';
 
 export default function PagingControl({
@@ -40,8 +40,10 @@ export default function PagingControl({
       <div style={styles.inlineFlex}>
         <Button
           onClick={changePage(-1)}
-          variant={'unstyled'}
-          isDisabled={pageNum == 0}
+          disabled={pageNum == 0}
+          marginRight="5px"
+          cursor="pointer"
+          unstyled
         >
           {PageSwitcherArrowLeft}
         </Button>
@@ -49,9 +51,11 @@ export default function PagingControl({
           {pageNum + 1} / {totalPages}
         </Text>
         <Button
-          variant={'unstyled'}
           onClick={changePage(1)}
-          isDisabled={pageNum == totalPages - 1}
+          disabled={pageNum == totalPages - 1}
+          marginLeft="5px"
+          cursor="pointer"
+          unstyled
         >
           {PageSwitcherArrowRight}
         </Button>
