@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '@web/context/AuthContext';
+import { AuthContext } from '@web/context/AuthContext.tsx';
 
 /**
  * @returns the authentication context wrapped in useContext
