@@ -20,8 +20,8 @@ function Review() {
       boxContent={
         <ReviewBox
           pdfFile={pdfFile}
-          name={formTemplateName ? formTemplateName : ''}
-          description={formTemplateDescription ? formTemplateDescription : ''}
+          name={formTemplateName ?? ''}
+          description={formTemplateDescription ?? ''}
           fieldGroups={fieldGroups}
         />
       }
@@ -34,4 +34,4 @@ function Review() {
   );
 }
 
-export default isAuth(Review, Scope.ADMIN);
+export default isAuth(Review, [Scope.ADMIN]);
