@@ -32,4 +32,4 @@ function Upload() {
   );
 }
 
-export default isAuth(Upload, Scope.ADMIN);
+export default isAuth(Upload, [Scope.ADMIN]);
